Model extended for heat overproduction and smaller corrections

+++ examples/memap/src/main/java/linprog/Simulation.java
@@ -11,7 +11,7 @@ public class Simulation {
 	private ActorTopology topology;
 
 	public static final int N_STEPS = 4*24; // 15 min timestep
-	public static final int N_DAYS = 1;
+	public static final double N_DAYS = 1; // 24 hours
 	
 	public static final int TIMESTEPS_PER_ITERATION = N_STEPS; //equals N_STEPS for case of overall optimization, smaller for reactive case
 	private static final int NR_OF_ITERATIONS = 1;
